use cluster rather than job-flow in remaining mrjob subcommands
(old commands and modules exist as deprecated aliases)
