+++ mrjob/tools/emr/create_job_flow.py
@@ -18,6 +18,7 @@ from sys import stderr
 
 from .create_cluster import main as real_main
 
+
 def main(args):
     print('create-job-flow is a deprecated alias for create-cluster'
           ' and will be removed in v0.6.0')
