+++ mrjob/tools/emr/terminate_job_flow.py
@@ -1,6 +1,4 @@
-# Copyright 2009-2012 Yelp
-# Copyright 2013 David Marin and Steve Johnson
-# Copyright 2015 Yelp
+# Copyright 2016 Yelp
 #
 # Licensed under the Apache License, Version 2.0 (the "License");
 # you may not use this file except in compliance with the License.
@@ -13,90 +11,18 @@
 # WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 # See the License for the specific language governing permissions and
 # limitations under the License.
-"""Terminate an existing EMR job flow.
+"""terminate_job_flow has been renamed to terminate_cluster; this is just
+a stub that runs terminate_cluster with a deprecation warning."""
+from __future__ import print_function
+from sys import stderr
 
-Usage::
-
-    mrjob terminate-job-flow [options] j-CLUSTERID
-    python -m mrjob.tools.emr.terminate_job_flow [options] j-CLUSTERID
-
-Terminate an existing EMR job flow.
-
-Options::
-
-  -h, --help            show this help message and exit
-  --aws-region=AWS_REGION
-                        Region to connect to S3 and EMR on (e.g. us-west-1).
-  -c CONF_PATHS, --conf-path=CONF_PATHS
-                        Path to alternate mrjob.conf file to read from
-  --no-conf             Don't load mrjob.conf even if it's available
-  --emr-endpoint=EMR_ENDPOINT
-                        Optional host to connect to when communicating with S3
-                        (e.g. us-west-1.elasticmapreduce.amazonaws.com).
-                        Default is to infer this from aws_region.
-  -q, --quiet           Don't print anything to stderr
-  --s3-endpoint=S3_ENDPOINT
-                        Host to connect to when communicating with S3 (e.g. s3
-                        -us-west-1.amazonaws.com). Default is to infer this
-                        from region (see --aws-region).
-  -t, --test            Don't actually delete any files; just log that we
-                        would
-  -v, --verbose         print more messages to stderr
-"""
-import logging
-from optparse import OptionParser
-
-from mrjob.emr import EMRJobRunner
-from mrjob.job import MRJob
-from mrjob.options import add_basic_opts
-from mrjob.options import add_emr_connect_opts
-from mrjob.options import alphabetize_options
-
-log = logging.getLogger(__name__)
+from .terminate_cluster import main as real_main
 
 
-def main(cl_args=None):
-    # parser command-line args
-    option_parser = make_option_parser()
-    options, args = option_parser.parse_args(cl_args)
-
-    if len(args) != 1:
-        option_parser.error('This tool takes exactly one argument.')
-    cluster_id = args[0]
-
-    MRJob.set_up_logging(quiet=options.quiet, verbose=options.verbose)
-
-    # create the persistent job
-    runner = EMRJobRunner(**runner_kwargs(options))
-    log.debug('Terminating job flow %s' % cluster_id)
-    runner.make_emr_conn().terminate_jobflow(cluster_id)
-    log.info('Terminated job flow %s' % cluster_id)
-
-
-def make_option_parser():
-    usage = '%prog [options] cluster-id'
-    description = 'Terminate an existing EMR job flow.'
-
-    option_parser = OptionParser(usage=usage, description=description)
-
-    option_parser.add_option(
-        '-t', '--test', dest='test', default=False,
-        action='store_true',
-        help="Don't actually delete any files; just log that we would")
-
-    add_basic_opts(option_parser)
-    add_emr_connect_opts(option_parser)
-    alphabetize_options(option_parser)
-
-    return option_parser
-
-
-def runner_kwargs(options):
-    kwargs = options.__dict__.copy()
-    for unused_arg in ('quiet', 'verbose', 'test'):
-        del kwargs[unused_arg]
-
-    return kwargs
+def main(args):
+    print('terminate-job-flow is a deprecated alias for'
+          ' terminate-cluster and will be removed in v0.6.0')
+    real_main(args)
 
 
 if __name__ == '__main__':
